requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
django-crispy-forms==1.10.0
djangorestframework==3.12.2
idna==2.10
Pygments==2.7.2
python-decouple==3.3
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
